2023-05-03 v1.1

+++ main.go
@@ -15,7 +15,7 @@ import (
 // 还是能用就行，不做错误检测
 
 // 全局变量
-var version string = `1.0`
+var version string = `1.1`
 var nginx_site_path, php_fpm_path string = `/etc/nginx/sites-enabled`, `unix:/var/run/php/php7.4-fpm.sock`
 
 // 初始化
@@ -51,31 +51,32 @@ func recover() {
 }
 
 // 带参数启动(暂未完善)
-// func etag() {
-// 	for i, v := range os.Args {
-// 		if i == 1 {
-// 			switch v {
-// 			case `uninstall`:
-// 				fmt.Println(`敬请期待！`)
-// 				os.Exit(0)
-// 			case `recover`:
-// 				recover()
-// 			case `clear`:
-// 				fmt.Println(`敬请期待！`)
-// 				os.Exit(0)
-// 			case `help`:
-// 				fmt.Println(`敬请期待！`)
-// 				os.Exit(0)
-// 			case `version`:
-// 				fmt.Println(version)
-// 				os.Exit(0)
-// 			default:
-// 				fmt.Printf("非法参数：%v\n", v)
-// 				os.Exit(0)
-// 			}
-// 		}
-// 	}
-// }
+func etag() {
+	for i, v := range os.Args {
+		if i == 1 {
+			switch v {
+			case `uninstall`:
+				fmt.Println(`敬请期待！`)
+				os.Exit(0)
+			case `recover`:
+				recover()
+			case `clear`:
+				fmt.Println(`敬请期待！`)
+				os.Exit(0)
+			case `help`:
+				fmt.Println(`敬请期待！`)
+				fmt.Println(`访问 https://github.com/ZxwyWebSite/NgConf 查看文档`)
+				os.Exit(0)
+			case `version`:
+				fmt.Println(version)
+				os.Exit(0)
+			default:
+				fmt.Printf("非法参数：%v\n", v)
+				os.Exit(0)
+			}
+		}
+	}
+}
 
 func head(path string) {
 	fmt.Print("\033[H\033[2J")
@@ -85,7 +86,7 @@ func head(path string) {
 }
 
 func main() {
-	//etag()
+	etag()
 	// 载入配置文件
 	cfg, err := ini.Load(`settings.ini`)
 	if err != nil {
@@ -100,7 +101,11 @@ func main() {
 		fmt.Println(`请检查settings.ini，如无数据，执行 "./ngconf recover" 重新初始化`)
 		os.Exit(1)
 	}
-	fmt.Println(`key: `, ver)
+	// 简单更新配置文件版本号
+	if ver.Value() != version {
+		cfg.Section(``).Key(`version`).SetValue(version)
+		cfg.SaveTo(`settings.ini`)
+	}
 	menu()
 }
 
@@ -144,7 +149,7 @@ func about() {
 	head(`关于程序`)
 	fmt.Println("\n当前版本：" + version)
 	fmt.Println(`Github：ZxwyWebSite/NgConf`)
-	fmt.Println(`专为海纳思系统开发，其它系统可能不兼容`)
+	fmt.Println(`专为海纳思系统开发，不兼容其它系统`)
 	os.Exit(0)
 }
 
@@ -174,7 +179,16 @@ func newsite() {
 		fmt.Print("\n\n是否确认生成配置(y|n): ")
 		fmt.Scanln(&confirm)
 		if confirm == `y` {
-			break
+			if site_path == `` {
+				fmt.Println("\n站点目录为空，请重新填写！")
+			} else {
+				break
+			}
+		} else {
+			if confirm == `n` {
+				fmt.Println("\n用户取消操作。")
+				os.Exit(0)
+			}
 		}
 	}
 	fmt.Println("\n开始生成配置文件，请稍候...")
@@ -210,7 +224,7 @@ func newsite() {
 	// 主配置文件 main.conf
 	main := `server
 {
-    # ipv6监听默认开启，如不需要可自行注释，https配置文件里还有一个
+    # ipv6监听(默认开启)，如不需要可自行注释，https配置文件里还有一个
     listen ` + site_port + `;
     listen [::]:` + site_port + `;
 	# 站点域名
@@ -227,14 +241,14 @@ func newsite() {
     #error_page 404 /404.html;
     #error_page 502 /502.html;
 
-    # 反向代理规则(默认关闭，需要修改)
-    # include ` + sitedir + `/proxy.conf` + `;
+    # URL重写(伪静态)配置(默认关闭，需要修改)
+    # include ` + sitedir + `/rewrite.conf` + `;
 
     # 启用PHP(默认开启)
     include ` + sitedir + `/php.conf` + `;
 
-    # URL重写(伪静态)配置(默认关闭，需要修改)
-    # include ` + sitedir + `/rewrite.conf` + `;
+    # 反向代理规则(默认关闭，需要修改)
+    # include ` + sitedir + `/proxy.conf` + `;
 
     #禁止访问的文件或目录
     location ~ ^/(\.user.ini|\.htaccess|\.git|\.env|\.svn|\.project|LICENSE|README.md)
@@ -257,33 +271,33 @@ func newsite() {
     error_log  ` + sitedir + `/error.log` + `;
 }`
 	// 伪静态 rewrite.conf
-	rewrite := `# # 二级目录填写 location /path，示例配置仅供参考，请根据实际情况修改
-# location / {
-#     if (!-e $request_filename){
-#         rewrite  ^(.*)$  /index.php$1  last;   break;
-#     }
-# }`
+	rewrite := `# 二级目录填写 location /path，示例配置仅供参考，请根据实际情况修改
+location / {
+    if (!-e $request_filename){
+        rewrite ^(.*)$ /index.php$1 last; break;
+    }
+}`
 	// 反向代理 proxy.conf
-	proxy := `# # 如需代理到二级目录可改为 location /path，示例配置仅供参考，请根据实际情况修改
-# location /
-# {
-#   # 要代理的网站域名
-#     proxy_pass https://translate.google.com;
-#   # 如果是内网|本机地址，填 $host
-#     proxy_set_header Host translate.google.com;
-#     proxy_set_header X-Real-IP $remote_addr;
-#     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-#     proxy_set_header Range $http_range;
-#   proxy_set_header If-Range $http_if_range;
-#     # proxy_set_header Cookie '你的cookie';
-#     proxy_http_version 1.1;
-#   # 字符串替换
-#     # proxy_set_header Accept-Encoding "";
-#   # sub_filter "Google 翻译" "Zxwy翻译";
-#     # sub_filter_once off;
-#   # 最大上传文件
-#     # client_max_body_size 20000m;
-# }`
+	proxy := `# 如需代理到二级目录可改为 location /path，示例配置仅供参考，请根据实际情况修改
+location /
+{
+    # 要代理的网站域名
+    proxy_pass https://translate.google.com;
+    # 如果是内网|本机地址，填 $host
+    proxy_set_header Host translate.google.com;
+    proxy_set_header X-Real-IP $remote_addr;
+    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    proxy_set_header Range $http_range;
+    proxy_set_header If-Range $http_if_range;
+    # proxy_set_header Cookie '你的cookie';
+    proxy_http_version 1.1;
+    # 字符串替换
+    # proxy_set_header Accept-Encoding "";
+    # sub_filter "Google 翻译" "Zxwy翻译";
+    # sub_filter_once off;
+    # 最大上传文件
+    # client_max_body_size 20000m;
+}`
 	// php设置 php.conf
 	php := `location ~ [^/]\.php(/|$)
 {
@@ -321,33 +335,51 @@ func newsite() {
     fastcgi_param SCRIPT_FILENAME $document_root$real_script_name;
     fastcgi_param SCRIPT_NAME $real_script_name;
     fastcgi_param PATH_INFO $path_info;
+    # fastcgi-php.conf
+    # regex to split $uri to $fastcgi_script_name and $fastcgi_path
+    #fastcgi_split_path_info ^(.+?\.php)(/.*)$;
+    # Check that the PHP script exists before passing it
+    #try_files $fastcgi_script_name =404;
+    # Bypass the fact that try_files resets $fastcgi_path_info
+    # see: http://trac.nginx.org/nginx/ticket/321
+    #set $path_info $fastcgi_path_info;
 }`
 	// 创建&写入配置文件
 	f1, _ := os.Create(sitedir + `/main.conf`)
-	defer f1.Close()
+	//defer f1.Close()
 	f1.Write([]byte(main))
+	f1.Close()
 	f2, _ := os.Create(sitedir + `/rewrite.conf`)
-	defer f2.Close()
+	//defer f2.Close()
 	f2.Write([]byte(rewrite))
+	f2.Close()
 	f3, _ := os.Create(sitedir + `/proxy.conf`)
-	defer f3.Close()
+	//defer f3.Close()
 	f3.Write([]byte(proxy))
+	f3.Close()
 	f4, _ := os.Create(sitedir + `/access.log`)
-	defer f4.Close()
+	//defer f4.Close()
+	f4.Close()
 	f5, _ := os.Create(sitedir + `/error.log`)
-	defer f5.Close()
+	//defer f5.Close()
+	f5.Close()
 	f6, _ := os.Create(sitedir + `/cert.pem`)
-	defer f6.Close()
+	//defer f6.Close()
+	f6.Close()
 	f7, _ := os.Create(sitedir + `/cert.key`)
-	defer f7.Close()
+	//defer f7.Close()
+	f7.Close()
 	f8, _ := os.Create(sitedir + `/php.conf`)
-	defer f8.Close()
+	//defer f8.Close()
 	f8.Write([]byte(php))
+	f8.Close()
 	f9, _ := os.Create(sitedir + `/ssl.conf`)
-	defer f9.Close()
+	//defer f9.Close()
 	f9.Write([]byte(ssl))
+	f9.Close()
 	fmt.Println("\n生成完毕，请到站点管理启用站点。")
 	fmt.Println("\n配置目录：" + sitedir)
+	fmt.Println(`如php程序有权限问题请执行 chown -R www-data:www-data ` + site_path)
 	os.Exit(0)
 }
 
@@ -356,12 +388,19 @@ func setsite() {
 	//fmt.Println(`敬请期待！`)
 	var siteinput int
 	var confirm string
-	var sitename [20]string
-	var siteactive [20]bool
 	fmt.Print("\n扫描中，请稍候...\n\n")
 	runpath, _ := os.Getwd()
 	sitedir := runpath + `/site`
 	files, _ := ioutil.ReadDir(sitedir)
+	// 获取文件夹数量，创建动态数组，解决上限20个站点问题
+	var sitenum int
+	for n := range files {
+		sitenum = n + 1
+	}
+	sitename := make([]string, sitenum)
+	siteactive := make([]bool, sitenum)
+	// var sitename [20]string
+	// var siteactive [20]bool
 	// 获取文件夹信息
 	for num, file := range files {
 		// 检测是否为目录
@@ -386,33 +425,39 @@ func setsite() {
 	//fmt.Println(`警告：这里也没有错误检测，请输入正确序号，然后二次确认`)
 	fmt.Print("\n请输入站点序号：")
 	fmt.Scanln(&siteinput)
-	if sitename[siteinput] != `` {
-		// 判断是否激活
-		if siteactive[siteinput] {
-			fmt.Println("\n将删除软链接 " + nginx_site_path + `/` + sitename[siteinput] + `.conf`)
-			fmt.Print("\n二次确认(y|n): ")
-			fmt.Scanln(&confirm)
-			if confirm == `y` {
-				os.Remove(nginx_site_path + `/` + sitename[siteinput] + `.conf`)
-				fmt.Println("\n设置完毕，请到Nginx状态重载配置。")
+	if siteinput < sitenum {
+		if sitename[siteinput] != `` {
+			// 判断是否激活
+			if siteactive[siteinput] {
+				fmt.Println("\n将删除软链接 " + nginx_site_path + `/` + sitename[siteinput] + `.conf`)
+				fmt.Print("\n二次确认(y|n): ")
+				fmt.Scanln(&confirm)
+				if confirm == `y` {
+					os.Remove(nginx_site_path + `/` + sitename[siteinput] + `.conf`)
+					fmt.Println("\n设置完毕，请到Nginx状态重载配置。")
+				} else {
+					fmt.Println("\n用户取消操作。")
+				}
 			} else {
-				fmt.Println("\n用户取消操作。")
+				fmt.Println("\n将创建软链接 " + sitedir + `/` + sitename[siteinput] + `/main.conf => ` + nginx_site_path + `/` + sitename[siteinput] + `.conf`)
+				fmt.Print("\n二次确认(y|n): ")
+				fmt.Scanln(&confirm)
+				if confirm == `y` {
+					os.Symlink(sitedir+`/`+sitename[siteinput]+`/main.conf`, nginx_site_path+`/`+sitename[siteinput]+`.conf`)
+					fmt.Println("\n设置完毕，请到Nginx状态重载配置。")
+				} else {
+					fmt.Println("\n用户取消操作。")
+				}
 			}
 		} else {
-			fmt.Println("\n将创建软链接 " + sitedir + `/` + sitename[siteinput] + `/main.conf => ` + nginx_site_path + `/` + sitename[siteinput] + `.conf`)
-			fmt.Print("\n二次确认(y|n): ")
-			fmt.Scanln(&confirm)
-			if confirm == `y` {
-				os.Symlink(sitedir+`/`+sitename[siteinput]+`/main.conf`, nginx_site_path+`/`+sitename[siteinput]+`.conf`)
-				fmt.Println("\n设置完毕，请到Nginx状态重载配置。")
-			} else {
-				fmt.Println("\n用户取消操作。")
-			}
+			fmt.Println("\n站点序号不存在。")
+			os.Exit(1)
 		}
 	} else {
 		fmt.Println("\n站点序号不存在。")
 		os.Exit(1)
 	}
+
 	// for n, f := range files {
 	// 	if n == siteinput {
 	// 		//fmt.Println(f.Name())
@@ -448,5 +493,6 @@ func nginx() {
 func uninstall() {
 	head(`卸载残留`)
 	fmt.Println(`敬请期待！`)
+	fmt.Println(`暂时可通过在站点管理中禁用站点解决`)
 	os.Exit(0)
 }
